P3: User can't stream HD on Android TV
Problem
When device connected to Ethernet, app would not choose hd video because its not connected to WiFi.

Solution
app choose by ConnectivityManager.isActiveNetworkMetered(). If it return false, app choose HD.

Explanation
A network is classified as metered when the user is sensitive to heavy data usage on that connection due to monetary costs, data limitations or battery/performance issues.

[#794302160783826]